fixed 1995 eof stop

+++ PITS_BLKEXIT_2016/programs/parser.h
@@ -9,5 +9,4 @@
 int check_atoi(int check, char* msg);
 int parse_input(char *str, int i, char *types[BUFFLEN], int limits[BUFFLEN], int precisions[BUFFLEN]);
 int read_specs(char *cfg_path, int col, char *types[BUFFLEN], int limits[BUFFLEN], int precisions[BUFFLEN]);
-long long int read_rows(char *cfg_path);
-
+int read_rows(char *cfg_path);
